requirements.txt
gradio
pillow
python-dotenv
opencv-python
numpy
groq
